+++ CHANGELOG.md
@@ -28,6 +28,9 @@ Additions
 - Added the ability to set the anchor that should be used when a tooltip attaches to an element or the mouse
 - Added the ability to display tooltips using the auto-nav style even when using the mouse
 
+Improvements
+- Allow scrolling panels to set height based on children with TreatSizeAsMaximum
+
 ## 6.3.1
 
 No code changes